% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av.R
\name{av}
\alias{av}
\title{av
Allows you to set "NULL" as an option in inputs. Shiny will recognize it as NULL}
\usage{
av(x)
}
\arguments{
\item{x}{input option}
}
\description{
av
Allows you to set "NULL" as an option in inputs. Shiny will recognize it as NULL
}
