% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbindPad.R
\name{cbindPad}
\alias{cbindPad}
\title{Add new column to a dataframe. If lengths differ, NA values will be used to fill.}
\usage{
cbindPad(...)
}
\description{
Add new column to a dataframe. If lengths differ, NA values will be used to fill.
}
