% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{data_sunmary
One line command to create summary of mean and SD}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{input identifier}

\item{varname}{y axis variable}

\item{groupnames}{vector of grouping variables to include}
}
\description{
data_sunmary
One line command to create summary of mean and SD
}
