% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPlot.R
\name{downloadPlot}
\alias{downloadPlot}
\title{One line command to create server options for plot download}
\usage{
downloadPlot(id, plotid)
}
\arguments{
\item{id}{input identifier}

\item{plotid}{name of the plot to be saved}
}
\description{
One line command to create server options for plot download
}
