% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPlotUI.R
\name{downloadPlotUI}
\alias{downloadPlotUI}
\title{One line command to create UI options for plot download}
\usage{
downloadPlotUI(id)
}
\arguments{
\item{id}{input identifier}
}
\description{
One line command to create UI options for plot download
}
