% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTable.R
\name{downloadTable}
\alias{downloadTable}
\title{One line command to create server options for table download}
\usage{
downloadTable(id, tableid)
}
\arguments{
\item{id}{input identifier}

\item{plotid}{name of the table to be saved}
}
\description{
One line command to create server options for table download
}
