% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTableUI.R
\name{downloadTableUI}
\alias{downloadTableUI}
\title{One line command to create UI options for table download}
\usage{
downloadTableUI(id)
}
\arguments{
\item{id}{input identifier}
}
\description{
One line command to create UI options for table download
}
