% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindPad.R
\name{rbindPad}
\alias{rbindPad}
\title{Allows you to add new rows of data to a column. Will automatically add NA values to rows in other columns.}
\usage{
rbindPad(data, selections)
}
\arguments{
\item{data}{Existing data table}

\item{selections}{vector of new elements to be added to column}
}
\description{
Allows you to add new rows of data to a column. Will automatically add NA values to rows in other columns.
}
